Removed public variable $this->duplicate_suffix in favour of manually setting it within the behavior.
Removed public variable $this->duplicate_suffix in favour of setting the variable manually each time the _deduplicate_slug method is called.

When creating multiple records from a script, this variable was not re-setting each time a new create() and then save() was called. I suspect that the behavior was initialised and the variable set only once which was at the point that the model was loaded.